+++ WidgetSets/TaharezLook/src/TLEditbox.cpp
@@ -87,7 +87,7 @@ TLEditbox::~TLEditbox(void)
 	Return the text code point index that is rendered closest to screen
 	position 'pt'.	
 *************************************************************************/
-ulong TLEditbox::getTextIndexFromPosition(const Point& pt) const
+size_t TLEditbox::getTextIndexFromPosition(const Point& pt) const
 {
 	//
 	// calculate final window position to be checked
