+++ WidgetSets/WindowsLook/include/WLEditbox.h
@@ -120,7 +120,7 @@ protected:
 	\return
 		Code point index into the text that is rendered closest to screen position \pt.
 	*/
-	virtual	ulong	getTextIndexFromPosition(const Point& pt) const;
+	virtual	size_t	getTextIndexFromPosition(const Point& pt) const;
 
 
 	/*!
